refactor: Validator 클래스 검증 로직 관련 도메인, DTO 클래스로 이동
