+++ src/main/java/racingcar/dto/RacingCarNamesRequest.java
@@ -1,7 +1,10 @@
 package racingcar.dto;
 
-import static racingcar.validator.Validator.*;
+import static racingcar.domain.Car.Validator.*;
+import static racingcar.dto.RacingCarNamesRequest.Validator.validateDuplicateCarNames;
+import static racingcar.exception.ExceptionMessage.DUPLICATE_CAR_NAME;
 
+import java.util.Arrays;
 import java.util.List;
 
 public class RacingCarNamesRequest {
@@ -23,6 +26,19 @@ public class RacingCarNamesRequest {
             validateCarNameLength(name);
         }
         validateDuplicateCarNames(names);
+    }
+
+    public static class Validator {
+        private Validator() {}
+
+        public static void validateDuplicateCarNames(String[] names) {
+            long noDuplicateCount = Arrays.stream(names)
+                    .distinct()
+                    .count();
+            if (noDuplicateCount != names.length) {
+                throw new IllegalArgumentException(DUPLICATE_CAR_NAME.getMessage());
+            }
+        }
     }
 
     public List<String> getNames() {
