--- src/main/java/racingcar/validator/Validator.java
@@ -1,46 +1,0 @@
-package racingcar.validator;
-
-import static racingcar.exception.ExceptionMessage.*;
-
-import java.util.Arrays;
-import java.util.List;
-import racingcar.domain.Car;
-
-public class Validator {
-    private static final int MAX_NAME_LENGTH = 6;
-
-    private Validator() {}
-
-    public static void validateNegativeTryCount(Integer tryCount) {
-        if (tryCount < 0) {
-            throw new IllegalArgumentException(ILLEGAL_TRY_COUNT.getMessage());
-        }
-    }
-
-    public static void validateDuplicateCarNames(String[] names) {
-        long noDuplicateCount = Arrays.stream(names)
-                .distinct()
-                .count();
-        if (noDuplicateCount != names.length) {
-            throw new IllegalArgumentException(DUPLICATE_CAR_NAME.getMessage());
-        }
-    }
-
-    public static void validateBlankCarName(String name) {
-        if(name.isBlank()) {
-            throw new IllegalArgumentException(BLANK_CAR_NAME.getMessage());
-        }
-    }
-
-    public static void validateCarNameLength(String name) {
-        if(name.length() >= MAX_NAME_LENGTH) {
-            throw new IllegalArgumentException(OUT_OF_CAR_NAME_LENGTH.getMessage());
-        }
-    }
-
-    public static void validateEmptyCars(List<Car> cars) {
-        if (cars.isEmpty()) {
-            throw new IllegalStateException(EMPTY_CARS.getMessage());
-        }
-    }
-}
